requirements.txt
args==0.1.0
astroid==2.3.3
bitrix24-rest==1.1.1
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
et-xmlfile==1.0.1
ftputil==3.4
idna==2.9
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
openpyxl==3.0.3
pylint==2.4.4
requests==2.23.0
six==1.14.0
urllib3==1.25.9
wrapt==1.11.2
